+++ ruma-events/src/room/member.rs
@@ -3,7 +3,7 @@
 use std::collections::BTreeMap;
 
 use ruma_events_macros::StateEventContent;
-use ruma_identifiers::{ServerKeyId, ServerNameBox, UserId};
+use ruma_identifiers::{ServerNameBox, ServerSigningKeyId, UserId};
 use serde::{Deserialize, Serialize};
 use strum::{Display, EnumString};
 
@@ -108,7 +108,7 @@ pub struct SignedContent {
 
     /// A single signature from the verifying server, in the format specified by the Signing Events
     /// section of the server-server API.
-    pub signatures: BTreeMap<ServerNameBox, BTreeMap<ServerKeyId, String>>,
+    pub signatures: BTreeMap<ServerNameBox, BTreeMap<ServerSigningKeyId, String>>,
 
     /// The token property of the containing third_party_invite object.
     pub token: String,
@@ -379,7 +379,6 @@ mod tests {
             "sender": "@alice:example.org",
             "state_key": "@alice:example.org"
         });
-
         assert_matches!(
             from_json_value::<Raw<StateEvent<MemberEventContent>>>(json)
                 .unwrap()