+++ ruma-identifiers/src/lib.rs
@@ -23,13 +23,11 @@ pub use crate::{
     room_id::RoomId,
     room_id_or_room_alias_id::RoomIdOrAliasId,
     room_version_id::RoomVersionId,
-    server_key_id::ServerKeyId,
     server_name::{ServerName, ServerNameBox},
     signatures::{
-        DeviceSignatures, DeviceSigningKeyId, EntitySignatures, KeyVersion, KeyVersionBox,
-        QualifiedKeyId, ServerSignatures, Signatures, SigningKeyId as SigningKeyTuple,
+        DeviceSignatures, DeviceSigningKeyId, EntitySignatures, KeyId, KeyName, KeyNameBox,
+        ServerSignatures, ServerSigningKeyId, Signatures, SigningKeyId,
     },
-    signing_key_id::SigningKeyId,
     user_id::UserId,
 };
 #[doc(inline)]
@@ -52,10 +50,8 @@ mod room_alias_id;
 mod room_id;
 mod room_id_or_room_alias_id;
 mod room_version_id;
-mod server_key_id;
 mod server_name;
 mod signatures;
-mod signing_key_id;
 
 /// Check whether a given string is a valid server name according to [the specification][].
 ///
@@ -141,7 +137,7 @@ macro_rules! room_version_id {
     };
 }
 
-/// Compile-time checked `ServerKeyId` construction.
+/// Compile-time checked `SigningKeyId` construction.
 #[macro_export]
 macro_rules! server_key_id {
     ($s:literal) => {