+++ ruma-identifiers/src/server_key_id.rs
@@ -1,2 +1,0 @@
-#[doc(inline)]
-pub type ServerKeyId = crate::signing_key_id::SigningKeyId;